+++ components/common/CharmEditor/utils.ts
@@ -41,18 +41,3 @@ export const safeRequestAnimationFrame =
         (window as any).lastTime = currTime + timeToCall;
         return id;
       };
-
-export function insertAndFocusFirstLine(view: EditorView) {
-  const { tr, doc, schema } = view.state;
-  const firstNode = doc.child(0);
-
-  if (!firstNode?.isText && !firstNode?.isTextblock) {
-    // create text node at the beginning of the document
-    const paragraph = schema.nodes.paragraph?.create();
-    tr.insert(0, paragraph);
-  }
-
-  // select the first line and focus
-  view.dispatch(tr.setSelection(TextSelection.atStart(tr.doc)));
-  view.focus();
-}
